Fixing compiler warnings

+++ drivers/input/touchscreen/oplus_touchscreen/touchpanel_healthinfo.c
@@ -5,6 +5,7 @@
 
 #include "touchpanel_healthinfo.h"
 
+#include <linux/types.h>
 #include <linux/err.h>
 #include <linux/string.h>
 #include <linux/sysfs.h>
@@ -1303,7 +1304,7 @@ int tp_healthinfo_report(void *tp_monitor_data, healthinfo_type type, void *valu
         ret = tp_grip_up_healthinfo_handle(monitor_data, *value_uint8, monitor_data->direction);
         break;
     case HEALTH_BELOW_RATE:
-        ret = tp_irq_interval_handle(monitor_data, (int)value);
+        ret = tp_irq_interval_handle(monitor_data, (uintptr_t)value);
         break;
     default:
         break;
